Integrate tqDist (#20)
Release candidate for v1.0.0.

* Prepare tqDist C code for R integration

* Update vignettes/clAnalyses.Rmd

Update per .pdf version

* tqdist Quartet Distance test trees

* tqdist TripletDistance test trees

* Tests of tqdist

* Preliminary wrapper functions for tqDist

* tqDist bindings

* Save tqDist bindings

* Integrate rtqdist

* document()

* Exporting triplet and quartet distance metrics

* Update tests/testthat/test-tqdist.R

C->R

* Missing header files

* Update src/rQuartetDist.cpp

}

* long long → int64_t

* long long -> int64_t

* long long → int64_t

* long long → int64_t

* unsigned long long → uint64_t

For portability

* NumericVector → IntegerVector

* Cast to int64_t

* Update .travis.yml

Use binary packages for faster install

* Suggest bookdown

* Binary packages not available

* Fewer random trees for Travis

* Rcpp Exception handling

* Don't include #include <cstdlib> // for exit

* #include <Rcpp.h>

* SlowQuartet → Quartet

* cerr / cout -> Rcpp::stop

* rm use.tqDist

* Whitespace

* Error message clarity

* Improve tests

* Spell out documentation

* Spacing

* Assert nDef doExtractAndContract

* Update src/HDT.cpp

+ return NULL;

* Update Quartet.Rproj

* tqdist_QuartetStatus

* tqdist_QuartetStatus

* Error checking

* tqdist_QuartetStatus

* tqdist_QuartetStatus

* Update man/MatchingQuartets.Rd

* Drop __declspec(dllexport)

* doExtractAndContract

* Update tests/testthat/test-tqdist.R

By default, returning maximum distance

* Faulty test for Unresolved trees

* Comments

* Hard-define quartetsToo

(informally)

* Brodal2013

* QuartetStatus return

* Comments from Brodal

* Spacing

* Consistent __ing

* Comments & whitespace

* Update NAMESPACE

* Remove commented verbose block

* (int) not required

* Count quartets resolved/unresolved in a single tree

* Use structure

* Move tqDist wrapper to tqDist.R

* Quartet statuses

* Update src/QuartetDistanceCalculator.cpp

Whitespace

* Document with Holt et al. 2014

* AllPairsQuartetStatus

* Numbers for sequencing

* Status → Agreement

* OneToMany quartet agreement

* Failure with E from oneToMany

* Use appropriate length

* Further test for 12many

* Don't need struct

* 2D not 3D

* Set size, rather than relying on push_back

* Go straight to IntegerVector

* Use integers in test

* Comparison tree as tree2

* MatchingQuartets now calculated directly by tqDist

* Move from Quartets to Partitions

* Rows, not columns

* Name rows by tree names

* Define bifurcators where needed

* Line length

* expect_identical (PQ)

* Format cf. MatchingQuartets

* Random trees are 1/3 similar

* Clean up tests

* Sequence update

* Update vignettes/clAnalyses.Rmd

Set Locale, to allow use of PCH 183

* Locale, for pch183

* Test cleanup

* document()

* Move raw data to separate repos

* CL data to CongreveLamsdell2016

* Create two_quartets.new

* Rotate matrices

* Load data from CongreveLamsdell2016 package

* Detail

* Remove O'Reilly datasets

Now hosted in OReillyEtAl2016 package on github

* Suggest devtools

* Use usethis

* Invert order

* Rows not cols

* Split out library installs

* require('CongreveLamsdell2016')

* Download from GitHub

* More inversion

* #TODOs

* Invert

* Update QuartetDistance.R

* Inversion

* Inversion

* Inversion

* +ghostscript

* Whitespace

* Further comment R code

* Update citation

* Update inst/CITATION

* Use library() and data()

* Drop forgotten ColSums call

* Data call

* Rename to QuartetStatus, SplitStatus

* +CongreveLamsdell2016

* t(docs)

* Examples

* Bipartition / Splits alias

* Aliases

* Delete clAnalyses.Rmd

Moved to CongreveLamsdell2016

* Export Bipartitions functions

* Update NAMESPACE

* document()

* Support single trees

* SharedQuartetStatus & SharedSplitStatus

* Exports

* Insert anonymised test trees

Which were previously causing trouble

* Revert "Insert anonymised test trees"

This reverts commit 3b984e5c1418cf559dce5c1dbd64c591f749104f.

* Debug test failures

* Move ms files to subdir

* Update ignores

* Replace preamble.tex

* Reduce size

* Indentation

* Check file exists

* Update cran-comments.md

* +Remotes

* Try removing integer casts

* Assume NO_N4_128

May not be as efficient, but at least it's (possibly) guaranteed to work.

* Complete final line

* SuppLEmentary files

(Ignore 'em)

* Schliep2011 (phangorn)

* Phangorn citation

* v0.2.0.9064

Celebrate 64-bit success!

* Refer to DOIs

* Input encoding & indentation

* R-hub failure on ubuntu

Phangorn-related

* Use doi, not URL

* Spelling

* Prepare for CRAN submission

* ISO c90 comment style

* Choices -> AllQuartets

* Improve documentation

* -ReleaseQuestions.R

* +Travis; +detail

* Documentation

* Documentation polish

* SplitsPoints → SplitPoints

* SplitPoints with unresolved comparison tree

* Improve test suite

* tqDist test suite coverage

* Improve test coverage

* Remove unicode character

+++ src/RootedTreeFactory.cpp
@@ -88,9 +88,9 @@ TemplatedLinkedList<RootedTree*>* RootedTreeFactory::getTemplatedLinkedList()
 	return returnMe;
 }
 
-long long RootedTreeFactory::getSizeInRam()
+int64_t RootedTreeFactory::getSizeInRam()
 {
-	long long resultRT = 0;
+  int64_t resultRT = 0;
 	{
 		RootedTree *current = createdRT;
 		while (current != NULL)
@@ -100,7 +100,7 @@ long long RootedTreeFactory::getSizeInRam()
 		}
 	}
 
-	long long resultTLL = 0;
+	int64_t resultTLL = 0;
 	{
 		TemplatedLinkedList<RootedTree*> *current = createdTLL;
 		while (current != NULL)
